split by files

+++ bitonic_sort.cl
@@ -1,6 +1,6 @@
 /**
  * Bitonic sort implementation.
  */
-__kernel void bitonicSort(__global TYPE* data, int sz) {
+__kernel void bitonicSort(__global TYPE* data, ulong sz) {
   uint gid = get_global_id(0);
 }
